% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_candidate.R
\name{jobdiva_create_candidate}
\alias{jobdiva_create_candidate}
\title{JobDiva Candidate Creation}
\usage{
jobdiva_create_candidate(
  first_name,
  last_name,
  title = "",
  company_name = "",
  email = "",
  alternate_email = "",
  work_phone = "",
  cell_phone = "",
  phx_employee_id = "",
  linkedin = "",
  city = "",
  state = ""
)
}
\arguments{
\item{first_name}{(type: string) --}

\item{last_name}{(type: string) --}

\item{company_name}{(type: string) --}

\item{email}{(type: string) --}

\item{alternate_email}{(type: string) --}

\item{phx_employee_id}{(type: string) -- The PHX Employee_Id of a candidate from Phoenix}

\item{linkedin}{(type: string) -- The LinkedIn URL of the candidate's LinkedIn profile}

\item{phone_numbers}{(type: string) -- The properly formatted string for phone number creation (action = 0
, phone = phone number, ext = "" and only for work phones, type: w = work, h = home, c = cell/mobile, f = fax, p = pager)}

\item{addresses}{(type: string) -- The properly formatted string for address creation (action = 0 for creation and action = 1 for updating)}
}
\value{
The JobDiva candidate id of newly created candidate
}
\description{
This function will insert information into JobDiva
}
