% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_candidate_bulk.R
\name{jobdiva_create_candidate_bulk}
\alias{jobdiva_create_candidate_bulk}
\title{JobDiva Bulk Candidate Creation}
\usage{
jobdiva_create_candidate_bulk(creation_df)
}
\arguments{
\item{creation_df}{(type: dataframe) -- a dataframe w/ the columns to create the candidate
(must at least contain first and last names otherwise the candidate will not be created. Currently does not handle phone numbers)}
}
\value{
A string that is the JobDiva candidate id of newly created candidate
}
\description{
This function will create candidates in JobDiva in bulk
}
