% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_candidate_hotlist.R
\name{jobdiva_create_candidate_hotlist}
\alias{jobdiva_create_candidate_hotlist}
\title{JobDiva Create Candidate Hotlist}
\usage{
jobdiva_create_candidate_hotlist(name, description = "", shared = c())
}
\arguments{
\item{name}{(type: string) -- a string indicating with the client id given by JobDiva}

\item{description}{(type: string) -- a string to describe the hotlist}

\item{shared}{(type: vector) -- a vector of users to share the hotlist with}
}
\value{
A string that is the id of the newly created hotlist.
}
\description{
This function will create a candidate hotlist in JobDiva and return new hotlist's id
}
