% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_candidate_note.R
\name{jobdiva_create_candidate_note}
\alias{jobdiva_create_candidate_note}
\title{JobDiva Create Candidate Note}
\usage{
jobdiva_create_candidate_note(
  candidate_id,
  note_body = "",
  note_owner_id = "",
  action = ""
)
}
\arguments{
\item{candidate_id}{(type: string) -- a string with the candidate id to add the note to}

\item{note_body}{(type: string) -- the body of the note}

\item{note_owner_id}{(type:string) -- the recruiter to be added as the note/activity owner}

\item{action}{(type: string) -- the note type (email, meeting, email opened, etc)}
}
\value{
(type: boolean) The function will return TRUE if the note was successfully created and FALSE otherwise.
}
\description{
This function will create a note for a candidate in JobDiva
}
