% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_company.R
\name{jobdiva_create_company}
\alias{jobdiva_create_company}
\title{JobDiva Create Company}
\usage{
jobdiva_create_company(
  company_name,
  company_website = "",
  phx_normalized_company_id = ""
)
}
\arguments{
\item{company_name}{(type: string) -- a string indicating with the client id given by JobDiva}

\item{company_website}{(type: string) -- email address of the designated API account}

\item{phx_normalized_company_id}{(type:string) -- the Phoenix Normalized Company Id to match the record to}
}
\value{
The JobDiva company id of the created company
}
\description{
This function will create a company in JobDiva if it does not already exist.
}
