% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_company_bulk.R
\name{jobdiva_create_company_bulk}
\alias{jobdiva_create_company_bulk}
\title{JobDiva Bulk Company Creation}
\usage{
jobdiva_create_company_bulk(creation_df)
}
\arguments{
\item{creation_df}{(type: dataframe) -- a dataframe w/ the columns to create the company (must at least contain company_name, but can also take company_website and normalized_company_id, columns MUST be named as such, but is case insensitive)}
}
\value{
A string that is the JobDiva contact id of newly created contact
}
\description{
This function will create companies in JobDiva in bulk
}
