% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_contact.R
\name{jobdiva_create_contact}
\alias{jobdiva_create_contact}
\title{JobDiva Contact Creation}
\usage{
jobdiva_create_contact(
  first_name,
  last_name,
  title = "",
  company_name = "",
  email = "",
  alternate_email = "",
  type = "",
  phone_numbers = "",
  phx_employee_id = "",
  linkedin = "",
  addresses = ""
)
}
\arguments{
\item{first_name}{(type: string) --}

\item{last_name}{(type: string) --}

\item{title}{(type: string) --}

\item{company_name}{(type: string) --}

\item{email}{(type: string) --}

\item{alternate_email}{(type: string) --}

\item{type}{(type: string) -- The type of contact}

\item{phone_numbers}{(type: string) -- The properly formatted string for phone number creation (action = 0
, phone = phone number, ext = "" and only for work phones, type: w = work, h = home, c = cell/mobile, f = fax, p = pager)}

\item{phx_employee_id}{(type: string) -- The PHX Employee_Id of a contact from Phoenix}

\item{linkedin}{(type: string) -- The LinkedIn URL of the contact's LinkedIn profile}

\item{addresses}{(type: string) -- The properly formatted string for address creation (action = 0 for creation and action = 1 for updating)}
}
\value{
The JobDiva contact id of newly created contact
}
\description{
This function will insert information into JobDiva
}
