% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_contact_bulk.R
\name{jobdiva_create_contact_bulk}
\alias{jobdiva_create_contact_bulk}
\title{JobDiva Bulk Contact Creation}
\usage{
jobdiva_create_contact_bulk(creation_df, type)
}
\arguments{
\item{creation_df}{(type: dataframe) -- a dataframe w/ the columns to create the contact
(must at least contain first and last names otherwise the contact will not be created. Currently does not handle phone numbers)}

\item{type}{(type: string) -- a string to indicate the contact type to associate the contact as (i.e. Hiring Manager)}
}
\value{
A string that is the JobDiva contact id of newly created contact
}
\description{
This function will create contacts in JobDiva in bulk
}
