% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_create_hotlist.R
\name{jobdiva_create_hotlist}
\alias{jobdiva_create_hotlist}
\title{JobDiva Create Hotlist}
\usage{
jobdiva_create_hotlist(
  name,
  active = TRUE,
  priv = FALSE,
  description = "",
  shared = c()
)
}
\arguments{
\item{name}{(type: string) -- a string indicating with the client id given by JobDiva}

\item{active}{(type: boolean) --  A boolean of whether or not the hotlist should be active}

\item{priv}{(type: boolean) -- A boolean of whether or not the hotlist should be private (only viewable to the person who created it/those who it is shared with)}

\item{description}{(type: string) -- a string to describe the hotlist}
}
\value{
A string that is the id of the newly created hotlist.
}
\description{
This function will create a hotlist in JobDiva and return new hotlist's id
}
