% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_deleted_records.R
\name{jobdiva_deleted_records}
\alias{jobdiva_deleted_records}
\title{JobDiva Deleted Records}
\usage{
jobdiva_deleted_records(
  entity,
  from_date = as.character(format(Sys.Date() - 1, "\%m/\%d/\%Y \%H:\%M:\%S")),
  to_date = as.character(format(Sys.Date(), "\%m/\%d/\%Y \%H:\%M:\%S"))
)
}
\arguments{
\item{entity}{(type: string) -- a string indicating which entity to find 
(Companies, Candidates, Contacts, or Jobs)}

\item{from_date}{(type: string) -- the earliest (starting) date for when the entity records where deleted (by default it is 1 day before the system date)}

\item{to_date}{(type: string) -- the lastest (ending) date for when the entity records where deleted (by default it is the system date)}
}
\value{
A dataframe of merged records based on the given timeframe and entity.
}
\description{
This function will return the dataframe that corresponds the recently deleted records for the given entity and timeframe.
}
