% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_detail.R
\name{jobdiva_detail}
\alias{jobdiva_detail}
\title{JobDiva Entity Detail}
\usage{
jobdiva_detail(
  entity,
  id_entity,
  clean_entity_df,
  udfs = "",
  bulk = TRUE,
  idcol = "id"
)
}
\arguments{
\item{entity}{(type: string) -- a string indicating which entity to find 
(Company, Candidate, Contact, etc.)}

\item{id_entity}{(type: string) --  what id are you using to identify the records
(i.e. for companynotesdetail: entity = 'companynotes', id_entity = 'company')}

\item{clean_entity_df}{(type: dataframe) -- a dataframe that must have a column named 'id' with the entity ids to query details for}

\item{udfs}{(type: vector) -- a vector of the UDFs to include}

\item{bulk}{(type: boolean) -- a boolean indicating whether we are upating 1 record or multiple}
}
\value{
A dataframe of entity details for the given entities.
}
\description{
This function will return the dataframe that corresponds the details for the entities submitted.
}
