% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_error_log.R
\name{jobdiva_error_log}
\alias{jobdiva_error_log}
\title{Report JobDiva API HTTP errors to Phoenix}
\usage{
jobdiva_error_log(r_function_name, http_response, phx_key, phx_user, path)
}
\arguments{
\item{r_function_name}{(type--string): The name of the function in R that called the API query.}

\item{http_response}{(type--list/response): The raw HTTP response directly from the query}

\item{phx_key}{(type-- Snowflake connector): The snowflake connector for login}

\item{phx_user}{(type-- string): The Phoenix username of the person using the function}

\item{path}{(type- string): The path to download temporary CSV files}
}
\value{
a Dataframe with the outcome of the error upload
}
\description{
This function will insert a record of an error from a JobDiva API query into the JobDiva_Import_Error_Log table in Phoenix.
}
