% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_get_emails.R
\name{jobdiva_get_emails}
\alias{jobdiva_get_emails}
\title{Query emails for a specific contact or candidate}
\usage{
jobdiva_get_emails(entity, entity_id)
}
\arguments{
\item{entity}{(type: string) -- a string indicating which entity to get 
(Candidate or Contact)}

\item{entity_id}{(type: string) -- Entity of the contact/candidate you want to query emails for}
}
\value{
Content from the HTTP response of the search
}
\description{
This function will return the HTTP response from get methods
in Nutshell
}
