% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_get_emails_bulk.R
\name{jobdiva_get_emails_bulk}
\alias{jobdiva_get_emails_bulk}
\title{Query emails for a dataframe of contacts or candidates}
\usage{
jobdiva_get_emails_bulk(creation_df, entity)
}
\arguments{
\item{entity}{(type: string) -- a string indicating which entity to get 
(Candidate or Contact)}

\item{entity_id}{(type: string) -- Entity of the contact/candidate you want to query emails for}
}
\value{
Content from the HTTP response of the search
}
\description{
This function will return the HTTP response from get methods
in Nutshell
}
