% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_get_saved_emails.R
\name{jobdiva_get_saved_emails}
\alias{jobdiva_get_saved_emails}
\title{JobDiva Saved Emails}
\usage{
jobdiva_get_saved_emails(
  from_date = as.character(format(Sys.Date() - 1, "\%m/\%d/\%Y \%H:\%M:\%S")),
  to_date = paste0(as.character(format(Sys.Date(), "\%m/\%d/\%Y")), " 23:59:59")
)
}
\arguments{
\item{from_date}{(type: string) -- the earliest (starting) date for when the entity records where updated (by default it is 1 day before the system date)}

\item{to_date}{(type: string) -- the lastest (ending) date for when the entity records where updated (by default it is the system date)}
}
\value{
A dataframe of new updated records based on the given timeframe.
}
\description{
This function will return the dataframe that corresponds the saved emails during the given timeframe.
}
