% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_hotlist_add_candidate_bulk.R
\name{jobdiva_hotlist_add_candidate_bulk}
\alias{jobdiva_hotlist_add_candidate_bulk}
\title{JobDiva Hotlist Add Candidate (bulk)}
\usage{
jobdiva_hotlist_add_candidate_bulk(jobdiva_candidate_id_vector, hotlist_id)
}
\arguments{
\item{jobdiva_candidate_id_vector}{(type: vector) -- The vector of candidate ids of candidates to be added to the given hotlist}

\item{hotlist_id}{(type: string) -- The id of the CANDIDATE hotlist that the candidates should be added to}
}
\value{
A dataframe of entity details for the given entities.
}
\description{
This function will bulk add candidates to a hotlist
}
