% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_hotlist_add_contact.R
\name{jobdiva_hotlist_add_contact}
\alias{jobdiva_hotlist_add_contact}
\title{JobDiva Hotlist Add Contact}
\usage{
jobdiva_hotlist_add_contact(hotlist_id, contact_id)
}
\value{
a status indicating either success or failure
}
\description{
This function will add the contact to the specified hotlist in JD.
}
