% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_hotlist_add_contact_bulk.R
\name{jobdiva_hotlist_add_contact_bulk}
\alias{jobdiva_hotlist_add_contact_bulk}
\title{JobDiva Hotlist Add Contact (bulk)}
\usage{
jobdiva_hotlist_add_contact_bulk(jobdiva_contact_id_vector, hotlist_id)
}
\arguments{
\item{entity}{(type: string) -- a string indicating which entity to find 
(Company, Candidate, Contact, etc.)}

\item{id_entity}{(type: string) --  what id are you using to identify the records
(i.e. for companynotesdetail: entity = 'companynotes', id_entity = 'company')}

\item{clean_entity_df}{(type: dataframe) -- a dataframe that must have a column named 'id' with the entity ids to query details for}

\item{udfs}{(type: vector) -- a vector of the UDFs to include}

\item{bulk}{(type: boolean) -- a boolean indicating whether we are upating 1 record or multiple}
}
\value{
A dataframe of entity details for the given entities.
}
\description{
This function will bulk add contacts to a hotlist
}
