% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_login_v2.R
\name{jobdiva_login_v2}
\alias{jobdiva_login_v2}
\title{JobDiva API Login v2}
\usage{
jobdiva_login_v2(
  client_id = 2425,
  username = "API.jobdiva@itmmi.com",
  pass = "ItmmI550$"
)
}
\arguments{
\item{client_id}{(type: string) -- a string indicating with the client id given by JobDiva}

\item{username}{(type: string) -- email address of the designated API account}

\item{pass}{(type: string) -- password of the designated API account}
}
\value{
A string that is the key for authentication into JobDiva's API
}
\description{
This function will return the key for authentication into JobDiva's API version 2
}
