% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_merge_company.R
\name{jobdiva_merge_company}
\alias{jobdiva_merge_company}
\title{JobDiva Merge Company}
\usage{
jobdiva_merge_company(dup_id, master_id)
}
\arguments{
\item{dup_id}{The Company id for the record to be merged into the main/master record}

\item{master_id}{The Company id for the record to inherit all of the notes and information from the dup record.}
}
\value{
a status indicating either success or failure
}
\description{
This function will merge the dup Company to the master Company (and all of their activity) in JD.
}
