% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_merge_contact.R
\name{jobdiva_merge_contact}
\alias{jobdiva_merge_contact}
\title{JobDiva Merge Contact}
\usage{
jobdiva_merge_contact(dup_id, master_id)
}
\arguments{
\item{dup_id}{The contact id for the record to be merged into the main/master record}

\item{master_id}{The contact id for the record to inherit all of the notes and information from the dup record.}
}
\value{
a status indicating either success or failure
}
\description{
This function will merge the dup contact to the master contact (and all of their activity) in JD.
}
