% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_search.R
\name{jobdiva_search}
\alias{jobdiva_search}
\title{Run search methods from JobDiva API}
\usage{
jobdiva_search(entity, offset = "")
}
\arguments{
\item{entity}{(type: string) -- a string indicating which entity to get 
(Company, Candidate,Contacts, etc.)}

\item{offset}{(type: integer) -- number of records to skip in query (useful for loops)}
}
\value{
Content from the HTTP response of the search
}
\description{
This function will return the HTTP response from search methods
in JobDiva
}
