% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_search_contact_by_company.R
\name{jobdiva_search_contact_by_company}
\alias{jobdiva_search_contact_by_company}
\title{Search Contact Profiles -- Company}
\usage{
jobdiva_search_contact_by_company(company)
}
\arguments{
\item{company}{(type: string) -- a string with the company name to query the contact by}
}
\value{
Content from the HTTP content of the search
}
\description{
This function will return the HTTP content from searching for a contact profile
}
