% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_search_contact_by_email.R
\name{jobdiva_search_contact_by_email}
\alias{jobdiva_search_contact_by_email}
\title{Search Contact Profiles}
\usage{
jobdiva_search_contact_by_email(contact_email)
}
\arguments{
\item{contact_email}{(type: string) -- a string with the email address to query the contact by}
}
\value{
Content from the HTTP content of the search
}
\description{
This function will return the HTTP content from searching for a contact profile
}
