% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_search_contact_by_nutshell_id.R
\name{jobdiva_search_contact_by_nutshell_id}
\alias{jobdiva_search_contact_by_nutshell_id}
\title{Search Contact Profiles}
\usage{
jobdiva_search_contact_by_nutshell_id(nutshell_id)
}
\arguments{
\item{contact_email}{(type: string) -- a string with the email address to query the contact by}
}
\value{
Content from the HTTP content of the search
}
\description{
This function will return the HTTP content from searching for a contact profile
}
