% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_udfs.R
\name{jobdiva_udfs}
\alias{jobdiva_udfs}
\title{Query User-Defined Fields (UDFs) from JobDiva}
\usage{
jobdiva_udfs()
}
\value{
a Dataframe with all UDFs in JobDiva
}
\description{
This function will return the list of UDFs in JobDiva
}
