% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_candidate_bulk.R
\name{jobdiva_update_candidate_bulk}
\alias{jobdiva_update_candidate_bulk}
\title{JobDiva Bulk Update candidates}
\usage{
jobdiva_update_candidate_bulk(update_df)
}
\arguments{
\item{update_df}{(type: dataframe) -- dataframe with the candidate information to be updated in JD. There must be a JobDiva_candidate_Id.}
}
\value{

}
\description{
This function will update candidate information in bulk in JobDiva
}
