% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_candidate_profile.R
\name{jobdiva_update_candidate_profile}
\alias{jobdiva_update_candidate_profile}
\title{JobDiva Update Candidate Profile}
\usage{
jobdiva_update_candidate_profile(jobdiva_candidate_id, update_df)
}
\arguments{
\item{jobdiva_candidate_id}{(type: string) -- a string with the JobDiva candidate id of the candidate
to be updated}

\item{update_df}{(type: dataframe) -- dataframe with the first column as 'FIELD' and second as 'CONTENT'}
}
\value{
status of whether or not the candidate was successfully updated.
}
\description{
This function will update standard candidate profile information in JobDiva
}
