% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_candidate_social_network_link.R
\name{jobdiva_update_candidate_social_link}
\alias{jobdiva_update_candidate_social_link}
\title{JobDiva Update Candidate Social Network Link}
\usage{
jobdiva_update_candidate_social_link(
  jobdiva_candidate_id,
  social_network_link,
  social_network_name
)
}
\arguments{
\item{jobdiva_candidate_id}{(type: string) -- a string with the JobDiva candidate id of the candidate
to be updated}

\item{update_df}{(type: dataframe) -- dataframe with the first column as 'FIELD' and second as 'CONTENT'}
}
\value{

}
\description{
This function will update candidate Social Network links information in JobDiva
}
