% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_company.R
\name{jobdiva_update_company}
\alias{jobdiva_update_company}
\title{JobDiva Update Company}
\usage{
jobdiva_update_company(jobdiva_company_id, update_df)
}
\arguments{
\item{jobdiva_company_id}{(type: string) -- a string with the JobDiva company id of the company 
to be updated}

\item{update_df}{(type: dataframe) -- dataframe with the first column as 'FIELD' and second as 'CONTENT'}
}
\value{

}
\description{
This function will update company information in JobDiva
}
