% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_company_bulk.R
\name{jobdiva_update_company_bulk}
\alias{jobdiva_update_company_bulk}
\title{JobDiva Bulk Update Companies}
\usage{
jobdiva_update_company_bulk(update_df)
}
\arguments{
\item{update_df}{(type: dataframe) -- dataframe with the company information to be updated in JD. There must be a JobDiva_Company_Id.}
}
\value{

}
\description{
This function will update company information in bulk in JobDiva
}
