% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_contact_bulk.R
\name{jobdiva_update_contact_bulk}
\alias{jobdiva_update_contact_bulk}
\title{JobDiva Bulk Update Contacts}
\usage{
jobdiva_update_contact_bulk(update_df)
}
\arguments{
\item{update_df}{(type: dataframe) -- dataframe with the contact information to be updated in JD. There must be a JobDiva_Contact_Id.}
}
\value{

}
\description{
This function will update contact information in bulk in JobDiva
}
