% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_contact_social_network_link.R
\name{jobdiva_update_contact_social_link_v2}
\alias{jobdiva_update_contact_social_link_v2}
\title{JobDiva Update Contact Social Network Link}
\usage{
jobdiva_update_contact_social_link_v2(
  jobdiva_contact_id,
  social_network_link,
  social_network_name
)
}
\arguments{
\item{jobdiva_contact_id}{(type: string) -- a string with the JobDiva contact id of the contact
to be updated}

\item{update_df}{(type: dataframe) -- dataframe with the first column as 'FIELD' and second as 'CONTENT'}
}
\value{

}
\description{
This function will update contact Social Network links information in JobDiva
}
