% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_phx_employee_id.R
\name{jobdiva_update_phx_employee_id}
\alias{jobdiva_update_phx_employee_id}
\title{JobDiva Update Contact PHX Employee id}
\usage{
jobdiva_update_phx_employee_id(jobdiva_contact_id, phx_employee_id)
}
\arguments{
\item{jobdiva_contact_id}{(type: string) -- The contact id}

\item{phx_employee_id}{(type: string) -- The PHX employee id to add to the record}
}
\value{
status of whether or not the contact's phx_employee_id was successfully updated.
}
\description{
This function will update contact PHX_Employee_ID in JobDiva
}
