% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_update_phx_normalized_company_id.R
\name{jobdiva_update_phx_normalized_company_id}
\alias{jobdiva_update_phx_normalized_company_id}
\title{JobDiva Update Company PHX Normalized Company id}
\usage{
jobdiva_update_phx_normalized_company_id(
  jobdiva_company_id,
  phx_normalized_company_id
)
}
\arguments{
\item{jobdiva_company_id}{(type: string) -- The company id}

\item{phx_normalized_company_id}{(type: string) -- The Normalized_Company id of the company in PHX to upload to JobDiva}
}
\value{
status of whether or not the company's phx_normalized_company_id was successfully updated.
}
\description{
This function will update contact PHX_Normalized_Company_ID in JobDiva
}
