% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_users.R
\name{jobdiva_users}
\alias{jobdiva_users}
\title{Query users from JD}
\usage{
jobdiva_users()
}
\value{
a Dataframe with users and their information from JobDiva
}
\description{
This function will return the list of users in JobDiva
}
