% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdiva_hotlist_add_contact_bulk.R
\name{jobdiva_hotlist_add_contact_bulk}
\alias{jobdiva_hotlist_add_contact_bulk}
\title{JobDiva Hotlist Add Contact (bulk)}
\usage{
jobdiva_hotlist_add_contact_bulk(jobdiva_contact_id_vector, hotlist_id)
}
\arguments{
\item{jobdiva_contact_id_vector}{(type: vector) -- The vector of contact ids of contacts to be added to the given hotlist}

\item{hotlist_id}{(type: string) -- The id of the CONTACT hotlist that the contacts should be added to}
}
\value{
A dataframe of entity details for the given entities.
}
\description{
This function will bulk add contacts to a hotlist
}
